Integrated gpt agents, gemini querier, gemini classifier

+++ application-manager/src/services/classifier/store/types/classifierResponseData.type.go
@@ -6,5 +6,5 @@ type ClassifierResponseData struct {
 
 type ClassData struct {
 	Name  string  `json:"name,omitempty" bson:"name,omitempty"`
-	Score float32 `json:"score,omitempty" bson:"score,omitempty"`
+	Score float64 `json:"score,omitempty" bson:"score,omitempty"`
 }
